Fold the buffer's helper types back into its own file
The reviewer's point on the previous commit stands: a struct whose only
purpose is to be one method's return type does not earn a file of its
own when that method is the single place it is built.

So the chunk iterators, the write-back guards, and the sub-word locator
move into the buffer's file, beside the methods that produce them. Every
one of them is now a private item: the four remaining constructors lose
their annotation, and the locator becomes private outright rather than
public within a private module. No visibility modifier is left in the
module, and none is needed.

The borrowed aliases stay where they are. They name the buffer rather
than being produced by it, callers spell them out directly, and nothing
in that file was ever annotated.
